o The Tests should be put into the same package as the original class   (Moved the ChapterDaoControllerIT into the same package within tests).
